fixed small bug and changed constructor from widht and height in grid units to meters

+++ bindings/python/pyObstacleGenerator.cpp
@@ -8,7 +8,7 @@ namespace py = pybind11;
 PYBIND11_MODULE(pyObstacleGenerator, m) {
 
     py::class_<ObstacleGenerator, ObstacleGenerator::Ptr>(m, "ObstacleGenerator")
-            .def(py::init<int, int, double>())
+            .def(py::init<double, double, double>())
             .def("run", &ObstacleGenerator::run)
             .def("setMaxObstacleNum", &ObstacleGenerator::setMaxObstacleNum)
             .def("setObstacleRadius", &ObstacleGenerator::setObstacleRadius)
